added authorization to Post model end points and tested it

+++ app/Http/Controllers/Api/PostController.php
@@ -7,6 +7,8 @@ use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
+use Illuminate\Support\Facades\Auth;
+
 
 class PostController extends Controller
 {
@@ -15,7 +17,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::with(['user:name','category:category_name'])->get();
+        $posts = Post::with(['user:id,name','category:id,category_name'])->get();
         $formantted_posts = $posts->map(function ($post){
             return [
                 'title' => $post->title,
@@ -33,8 +35,7 @@ class PostController extends Controller
             ];
         });
         return response()->json($formantted_posts,200);
-
-        }
+    }
 
     /**
      * Store a newly created resource in storage.
@@ -46,16 +47,14 @@ class PostController extends Controller
             [
                 'title' => 'required|string',
                 'post' => 'required|string',
-                'category_id' =>'required|int',
-                'user_id' => 'required|int'
+                'category_id' =>'required|int|exists:categories,id',
+                
             ],
             [
                 'title.required' => 'title is required',
                 'post.required' => 'post is required',
                 'category_id.required' => 'choose category for the post',
                 'category_id.exists' => 'the category you chose does not exist',
-                'user_id.required' => 'the use id is required',
-                'user_id.exists' => 'this user is not registered'
             ]
            );
         }
@@ -66,11 +65,15 @@ class PostController extends Controller
                     'erros' => $e->errors()
                 ], 422);
         }
-           Post::create($validated + ['status' => 'created']);
+         $post =  Post::create($validated + [
+           'user_id' => $request->user()->id, 
+           'status' => 'created'
+           ]);
         return response()->json([
+            'post' => $post->id,
             'message' => 'post created',
             
-        ],200);
+        ],201);
         
     }
 
@@ -79,7 +82,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        $post = $post->load(['user:name,id','categories:category_name'])->get();
+        $post = $post->load(['user:id,name','category:id,category_name']);
         return response()->json(
             [
                 'title' => $post->title,
@@ -90,7 +93,7 @@ class PostController extends Controller
                 ],
                 'category' => [
                 'category_id' => $post->category_id,
-                'category_name' => $post->category_name
+                'category_name' => $post->category->category_name
             ]
             ]
             ,200);
@@ -102,46 +105,45 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        try{
+        if (! $request->user()->id == $post->user_id){
+            return response()->json([
+
+                'message' => 'unauthorized!'
+            ],403);
+        }
            $validated = $request->validate(
             [
                 'title' => 'required|string',
                 'post' => 'required|string',
-                'category_id' =>'required|int|exists:posts,category_id',
-                'user_id' => 'required|int|exists:posts,user_id'
+                'category_id' =>'required|int|exists:categories,id',
             ],
             [
                 'title.required' => 'title is required',
                 'post.required' => 'post is required',
                 'category_id.required' => 'choose category for the post',
                 'category_id.exists' => 'the category you chose does not exist',
-                'user_id.required' => 'the use id is required',
-                'user_id.exists' => 'this user is not registered'
             ]
            );
-        }
-        catch(ValidationException $e){
-            return response()->json(
-                [
-                    'message' => 'validation failed',
-                    'erros' => $e->errors()
-                ], 422);
-        }
-        $post->update($validated + ['status' => 'edited']);
-        return Response()->json([
+        $post = $post->update($validated + ['status' => 'edited']);
+        return response()->json([
                 'message' => 'post edited!'
-        ]
-        );
+        ],201);
     }
 
         //
-    public function destroy(Post $post)
+    public function delete(Post $post)
     {
+        if (! Auth::user()->id == $post->user_id || Auth::user()->role == 'admin'){
+            return response()->json([
+
+                'message' => 'unauthorized!'
+            ],403);
+        }
         $post->delete();
         return response()->json(
             [
                 'message' => 'post deleted!'
             ],
-        204);
+        201);
     }
 }
